requirements.txt
uvloop==0.14.0
uvicorn==0.13.3
aiofiles==0.6.0
fastapi==0.63.0
gunicorn==20.0.4
httptools==0.1.1
loguru==0.5.3
Pillow==8.1.0
pydantic==1.7.3
python-multipart==0.0.5
h11==0.12.0
starlette==0.13.6